refactor: move ourValues interface to component file and remove comments
